app/requirements.txt
attrs==25.1.0
Automat==24.8.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
cssselect2==0.7.0
defusedxml==0.7.1
filelock==3.16.1
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
packaging==24.2
parsel==1.9.1
pillow==10.4.0
Protego==0.3.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
queuelib==1.7.0
reportlab==4.3.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.2.0
SQLAlchemy==2.0.38
svglib==1.5.1
tinycss2==1.4.0
tldextract==5.1.2
tomli==2.2.1
tqdm==4.67.1
twisted==24.11.0
typing-extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
webencodings==0.5.1
zope.interface==7.2